paralleled computations of active paths and metrics
