+++ src/core/convergence.py
@@ -3,6 +3,8 @@ from dataclasses import dataclass
 from math import ceil
 from typing import Dict, List, Tuple, Collection, Callable
 
+
+from core.active_paths import PathsOverTime, get_activity_indicator, compute_path_travel_time, compute_all_active_paths, all_active_paths_parallel
 from core.bellman_ford import bellman_ford
 from core.dijkstra import reverse_dijkstra
 from core.dynamic_flow import DynamicFlow
@@ -19,51 +21,9 @@ from utilities.arrays import elem_lrank
 from visualization.to_json import merge_commodities, to_visualization_json
 
 from concurrent.futures import ThreadPoolExecutor
+import multiprocessing as mp
 
 Path = List[Edge]
-
-#
-# @dataclass
-# class LabelledPathEntry:
-#     """
-#     This class represents one edge of a labelled path.
-#     Each such edge is labelled with a starting time and an active_until time.
-#     """
-#
-#     edge: Edge
-#     starting_time: float
-#     active_until: float
-
-
-class PathsOverTime:
-    """
-    Collection of paths with same source and sink.
-    Each path is equipped with intervals of activity represented by indicator
-    """
-
-    paths: List[Path]
-    activity_indicators: List[Indicator]
-
-    def __init__(self, paths, activity_indicators):
-        self.paths = paths
-        self.activity_indicators = activity_indicators
-
-    def __len__(self):
-        return len(self.paths)
-
-    def __add__(self, other):
-        return PathsOverTime(
-            self.paths + other.paths,
-            self.activity_indicators + other.activity_indicators
-        )
-
-    def add_path(self, path, activity_indicator):
-        if path in self.paths:
-            i = self.paths.index(path)
-            self.activity_indicators[i] += activity_indicator
-        else:
-            self.paths += [path]
-            self.activity_indicators += [activity_indicator]
 
 
 def integrate_with_weights(lin: PiecewiseLinear, weights: RightConstant, start: float, end: float):
@@ -101,7 +61,7 @@ class BaseFlowIterator:
     _comm_to_path: Dict[int, Path]
     _path_to_comm: Dict[Tuple, int]
     _paths: Dict[Tuple[Node, Node], List[Path]]  # all used s-t paths
-    _path_metrics: Dict[Tuple[Node, Node], List[List[Tuple[Path, Dict]]]]
+    _path_metrics: Dict[Tuple[int, int], Dict[int, Tuple]]
     _important_nodes: Dict[Tuple[Node, Node], Collection[Node]]
     _fastest_travel_times: Dict[Tuple[Node, Node], float]  # travel times without queues
     _earliest_arrivals_to: Dict[Node, Dict[Node, PiecewiseLinear]] # earliest arrivals to sinks
@@ -136,7 +96,7 @@ class BaseFlowIterator:
             s, inflow = next(iter(com.sources.items()))
             t = com.sink
             self._paths[(s, t)] = []
-            self._path_metrics[(s, t)] = []
+            self._path_metrics[(s.id, t.id)] = dict()
             self._important_nodes[(s, t)] = (
                     network.graph.get_nodes_reaching(t) & network.graph.get_reachable_nodes(s)
             )
@@ -265,40 +225,6 @@ class BaseFlowIterator:
 
         return avg_delays
 
-    def _compute_path_metrics(self, path):
-        metrics = dict()
-
-        s = path[0]._node_from
-        t = path[-1]._node_to
-
-        costs = self._flow.get_edge_costs()
-        optimal_travel_time = self._earliest_arrivals_to[t][s] - identity
-
-        path_inflow = self._get_path_inflow(path).simplify()
-        accum_path_inflow = path_inflow.integral()
-        path_delay = compute_path_travel_time(path, costs) - optimal_travel_time
-
-        metrics['avg_delay_normalized'] = (
-                integrate_with_weights(path_delay, path_inflow, 0, self.inflow_horizon) /
-                (accum_path_inflow(self.inflow_horizon) * self._fastest_travel_times[(s, t)])
-        )
-
-        metrics['share_of_total_inflow'] = (
-                accum_path_inflow(self.inflow_horizon) /
-                self._inflows[(s, t)].integral()(self.inflow_horizon)
-        )
-
-        activity_ind = get_activity_indicator(
-            path_delay,
-            threshold=self.delay_threshold * self._fastest_travel_times[(s, t)]
-        )
-        metrics['active_inflow_share'] = (
-                (path_inflow * activity_ind).integral()(self.inflow_horizon) /
-                accum_path_inflow(self.inflow_horizon)
-        )
-
-        return metrics
-
     def run(self, num_iterations, eval_every=10):
         """
         Main cycle
@@ -328,10 +254,24 @@ class BaseFlowIterator:
                 print(f"Normalized average delays:")
                 for (s, t), avg_delay in self._get_route_avg_delays(normalize=True).items():
                     print(f"({s.id} -> {t.id}): {round(avg_delay, 4)}")
-                    self._path_metrics[(s, t)].append([(path, self._compute_path_metrics(path)) for path in self._paths[(s, t)]])
+
+                for (s, t) in self._paths.keys():
+                    network_data = (
+                        self._flow.get_edge_costs(),
+                        self._earliest_arrivals_to[t][s] - identity,
+                        self._fastest_travel_times[(s, t)],
+                        self._inflows[(s, t)],
+                        self.inflow_horizon,
+                        self.delay_threshold
+                    )
+
+                    with mp.Pool(processes=mp.cpu_count()) as pool:
+                        self._path_metrics[(s.id, t.id)][self._iter] = pool.starmap(
+                             compute_path_metrics,
+                             [(path, self._get_path_inflow(path), network_data) for path in self._paths[(s, t)]]
+                        )
 
         merged_flow = self._flow
-        # combine_commodities_with_same_sink(self.network)
         for paths in self._paths.values():
             commodities = [self._path_to_comm[tuple(p)] for p in paths]
             merged_flow = merge_commodities(merged_flow, self.network, commodities)
@@ -339,7 +279,36 @@ class BaseFlowIterator:
         return merged_flow, self._path_metrics
 
 
-# class AlphaFlowIterator(BaseFlowIterator):
+def compute_path_metrics(path, path_inflow, network_data):
+
+    costs, optimal_travel_time, fastest_travel_time, total_inflow, inflow_horizon, delay_threshold = network_data
+
+    metrics = {'path': [e.id for e in path]}
+
+    accum_path_inflow = path_inflow.integral()
+    path_delay = compute_path_travel_time(path, costs) - optimal_travel_time
+    activity_ind = get_activity_indicator(
+        path_delay,
+        threshold=delay_threshold * fastest_travel_time
+    )
+
+    metrics['avg_delay_normalized'] = (
+            integrate_with_weights(path_delay, path_inflow, 0, inflow_horizon) /
+            (accum_path_inflow(inflow_horizon) * fastest_travel_time)
+    )
+
+    metrics['share_of_total_inflow'] = (
+            accum_path_inflow(inflow_horizon) /
+            total_inflow.integral()(inflow_horizon)
+    )
+    metrics['active_inflow_share'] = (
+            (path_inflow * activity_ind).integral()(inflow_horizon) /
+            accum_path_inflow(inflow_horizon)
+    )
+
+    return metrics
+
+# class AlphaFlowIterator(BaseFlowIterator): old version, uncommenting will probably result in errors
 #     alpha: float    # fraction of inflow to be redistributed
 #     approx_inflows: bool    # option to project inflows to regular grid after each iteration
 #     """
@@ -420,134 +389,6 @@ class BaseFlowIterator:
 #         #     process_route(route)
 
 
-# def compute_shortest_paths_over_time(
-#     earliest_arrivals: Dict[Node, PiecewiseLinear],
-#     edge_costs: List[PiecewiseLinear],
-#     source: Node,
-#     sink: Node,
-# ) -> PathsOverTime:
-#     """
-#     Returns the shortest paths from source to sink in the network over time.
-#     """
-#     shortest_paths_computed_until = 0.0
-#     identity = PiecewiseLinear(
-#         [shortest_paths_computed_until],
-#         [shortest_paths_computed_until],
-#         1.0,
-#         1.0,
-#         (shortest_paths_computed_until, float("inf")),
-#     )
-#     edge_exit_times: Dict[Edge, PiecewiseLinear] = {}
-#
-#     paths_over_time: PathsOverTime = PathsOverTime([], [])
-#
-#     while shortest_paths_computed_until < float("inf"):
-#         labelled_path: List[LabelledPathEntry] = []
-#
-#         path_start = departure = shortest_paths_computed_until
-#         # We want to find a path that is active starting from time `departure` and that is active for as long as possible (heuristically?).
-#
-#         # We start at the source and iteratively select the next edge of the path.
-#         v = source
-#         while v != sink:
-#             # Select the next outgoing edge of the current node v:
-#             # the edge e that is active for the longest period (starting from the arrival/departure time at v `departure`).
-#             best_edge = None
-#             best_active_until = None
-#
-#             for edge in v.outgoing_edges:
-#                 edge_exit_times[edge] = identity.plus(
-#                     edge_costs[edge.id]
-#                 ).ensure_monotone(True)
-#                 edge_delay = (
-#                         earliest_arrivals[edge.node_to].compose(edge_exit_times[edge])
-#                         - earliest_arrivals[v]
-#                 ).simplify()
-#
-#                 # it can happen that delay is negative - just computational issue?
-#
-#                 # edge_delay is (close to) zero at times when the edge is active; otherwise it is positive.
-#                 if edge_delay(departure) > eps:
-#                     continue
-#
-#                 active_until = edge_delay.next_change_time(departure)
-#
-#                 if best_edge is None or active_until > best_active_until:
-#                     best_edge = edge
-#                     best_active_until = active_until
-#
-#             assert best_edge is not None
-#             labelled_path.append(
-#                 LabelledPathEntry(best_edge, departure, best_active_until)
-#             )
-#             v = best_edge.node_to
-#             departure = edge_exit_times[best_edge](departure)
-#
-#         # Compute path_active_until
-#         rest_of_path_active_until = float("inf")
-#         for labelled_edge in reversed(labelled_path):
-#             last_enter_time_st_rest_path_active = edge_exit_times[
-#                 labelled_edge.edge
-#             ].max_t_below(rest_of_path_active_until)
-#             rest_of_path_active_until = min(
-#                 last_enter_time_st_rest_path_active, labelled_edge.active_until
-#             )
-#
-#         paths_over_time.add_path(
-#             [label.edge for label in labelled_path],
-#             Indicator.from_interval(path_start, rest_of_path_active_until)
-#         )
-#
-#         assert shortest_paths_computed_until < rest_of_path_active_until
-#
-#         shortest_paths_computed_until = rest_of_path_active_until
-#     return paths_over_time
-
-
-def get_activity_indicator(
-        delay: PiecewiseLinear,
-        threshold: float = eps,
-        min_interval_length: float = eps
-) -> Indicator:
-    """
-    Returns function I(t) with value 1 if delay(t) < threshold and 0 otherwise.
-    Intervals of length less than min_interval_length are avoided.
-    """
-
-    times = []
-    values = []
-
-    is_active = False
-    for i in range(len(delay.times)):
-        if delay.values[i] < threshold and not is_active:
-            if i > 0:
-                offset = (delay.values[i] - threshold) / delay.gradient(i - 1)
-                interval_start = delay.times[i] - offset
-            else:
-                interval_start = delay.domain[0]
-
-            is_active = True
-        elif delay.values[i] > threshold and is_active:
-            offset = (threshold - delay.values[i - 1]) / delay.gradient(i - 1)
-            interval_end = delay.times[i-1] + offset
-
-            is_active = False
-
-            if interval_end - interval_start > min_interval_length:
-                times += [interval_start, interval_end]
-                values += [1.0, 0.0]
-
-    if is_active:  # assuming last slope for delay can only equal to 0
-        times += [interval_start]
-        values += [1.0]
-
-    if len(times) == 0 or times[0] > delay.domain[0] + eps:
-        times = [delay.domain[0]] + times
-        values = [0.0] + values
-
-    return Indicator(times, values, delay.domain)
-
-
 def approximate_linear(lin: PiecewiseLinear, delta: float, horizon: float) -> RightConstant:
     """
     Returns a RightConstant approximation of a given PiecewiseLinear with grid size delta
@@ -560,68 +401,6 @@ def approximate_linear(lin: PiecewiseLinear, delta: float, horizon: float) -> Ri
     new_values[-1] = lin.values[-1]
 
     return RightConstant(new_times, new_values, lin.domain)
-
-
-def compute_path_travel_time(path: Path, costs: List[PiecewiseLinear]) -> PiecewiseLinear:
-    path_exit_time = identity
-
-    for edge in path[::-1]:
-        path_exit_time = path_exit_time.compose(
-            identity.plus(costs[edge.id]).ensure_monotone(True)
-        )
-
-    return path_exit_time - identity
-
-
-def compute_all_active_paths(
-    earliest_arrival_fcts: Dict[Node, PiecewiseLinear],
-    edge_costs: List[PiecewiseLinear],
-    source: Node,
-    sink: Node,
-    inflow_horizon: float,
-    delay_threshold: float = eps,
-    min_active_time: float = 1000*eps
-) -> PathsOverTime:
-    """
-    Constructs all simple s-t paths which are active for longer than min_active_time
-    """
-
-    shortest_paths = PathsOverTime([], [])
-
-    initial_paths = PathsOverTime(
-            paths=[[]],
-            activity_indicators=[Indicator.from_interval(0.0, float('inf'))]
-        )
-    paths_to = {source: (initial_paths, [identity])}  # stores constructed subpaths grouped by end node
-
-    while len(paths_to) > 0:
-        destinations = set(e._node_to for v in paths_to.keys() for e in v.outgoing_edges if v != sink)
-        extended_paths_to = {
-            v: (PathsOverTime([], []), [])
-            for v in destinations
-        }
-
-        for v, (paths_to_v, exit_times) in paths_to.items():
-            if v == sink:
-                shortest_paths += paths_to_v
-                continue
-            for e in v.outgoing_edges:
-                for i in range(len(paths_to_v)):  # extend each path by adding an edge
-                    if e in paths_to_v.paths[i]:    # do not allow loops
-                        continue
-                    path = paths_to_v.paths[i] + [e]
-                    exit_time = (identity + edge_costs[e.id]).compose(exit_times[i])
-                    delay = earliest_arrival_fcts[e._node_to].compose(exit_time) - earliest_arrival_fcts[source]
-                    indicator = get_activity_indicator(delay, threshold=delay_threshold)
-                    if indicator.integral()(inflow_horizon) > min_active_time:
-                        extended_paths_to[e._node_to][0].add_path(path, indicator)
-                        extended_paths_to[e._node_to][1].append(exit_time)
-
-        paths_to = {v: (paths_to_v, exit_times)
-                    for v, (paths_to_v, exit_times) in extended_paths_to.items()
-                    if len(paths_to_v) > 0}
-
-    return shortest_paths
 
 
 class AlphaFlowIterator(BaseFlowIterator):
@@ -692,7 +471,7 @@ class AlphaFlowIterator(BaseFlowIterator):
 
         def process_route(route):
             s, t = route
-            active_paths = compute_all_active_paths(
+            active_paths = all_active_paths_parallel(
                 self._earliest_arrivals_to[t],
                 costs,
                 s,
@@ -709,8 +488,8 @@ class AlphaFlowIterator(BaseFlowIterator):
                     approximation = inflow.project_to_grid(self.reroute_interval, self.inflow_horizon)
                     self._set_path_inflow(path, approximation)
 
-        for route in self._paths.keys():
-            process_route(route)
+        # for route in self._paths.keys():
+        #     process_route(route)
 
-        # with ThreadPoolExecutor() as executor:
-        #     executor.map(process_route, self._paths.keys())
+        with ThreadPoolExecutor() as executor:
+            executor.map(process_route, self._paths.keys())
