+++ src/core/graph.py
@@ -39,6 +39,12 @@ class Node:
         self._outgoing_edges = []
         self._graph = graph
 
+    def __eq__(self, other):
+        return self.id == other.id
+
+    def __hash__(self):
+        return hash(self.id)
+
     @property
     def incoming_edges(self):
         return (
